退货

+++ jcart-store-back/src/main/java/io/cxy/jcartstoreback/po/Return.java
@@ -5,7 +5,7 @@ import java.util.Date;
 public class Return {
     private Integer returnId;
 
-    private Long orderId;
+    private Integer orderId;
 
     private Date orderTime;
 
@@ -45,11 +45,11 @@ public class Return {
         this.returnId = returnId;
     }
 
-    public Long getOrderId() {
+    public Integer getOrderId() {
         return orderId;
     }
 
-    public void setOrderId(Long orderId) {
+    public void setOrderId(Integer orderId) {
         this.orderId = orderId;
     }
